chore: add hunts and rework profile routes

+++ src/api/profiles/models.rs
@@ -7,7 +7,7 @@ pub struct UserProfiles {
     pub id: Uuid,
     pub user_id: Uuid,
     pub points: i32,
-    pub level: i32,
+    pub level: f32,
     pub completed_hunts: i32,
     pub updated_at: DateTime<Utc>,
 }
